refactor(dashboard): extract Mastio tools + network sub-router (F-B-201 PR-5/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-201.md

Sprint F-B-201 PR-5 of 10. base=main. Combined into one sub-router
because the tools and network surfaces share no behaviour with the
rest of the router and the network discovery page is the natural
sibling of the tool-registry viewer (both are read-only directory
browsers, plus one CSRF-gated reload).

Extracts 3 routes (~130 LOC) from mcp_proxy/dashboard/router.py.

  mcp_proxy/dashboard/tools_network_routes.py (new, 166 LOC, 3 routes):
    GET  /proxy/tools          tool registry list (read-only)
    GET  /proxy/network        cross-org agent discovery directory
    POST /proxy/tools/reload   reload YAML tool config (CSRF-gated)

  mcp_proxy/dashboard/router.py (3720 -> 3592 LOC, -128 net):
    - 3 route bodies removed; breadcrumb at the old position.
    - One new include_router(_tools_network_routes.router) call.

Behavioural delta: zero. /tools/reload CSRF check preserved.
/network bridge-lookup + filter pipeline preserved (querier default,
include_own_org default, error surfaces).

Tests:
  - pytest tests/test_proxy_dashboard_oidc.py
    tests/test_proxy_dashboard_mcp_resources.py
    tests/test_dashboard.py tests/test_dashboard_principals.py
    tests/test_h9_login_lockout.py
    -n auto --dist=loadfile -q
  - 100 passed.

Router LOC trajectory:
  Pre-Sprint:    66 routes, 5106 LOC
  Post-PR-1 #859: 66 routes, 5052 LOC
  Post-PR-2 #865: 61 routes, 4828 LOC
  Post-PR-3 #866: 58 routes, 4117 LOC
  Post-PR-4 #868: 51 routes, 3720 LOC
  Post-PR-5 (this): 48 routes, 3592 LOC
  Target post-PR-10: ~3 routes, ~120 LOC aggregator.

Next PR-6 (F-B-201 modularization): /proxy/policies surface
(page + save + test-webhook, 3 routes ~180 LOC) into policies_routes.py.
